[codex] Organize concurrent runtime internals
