+++ src/internal/runFork.ts
@@ -1,13 +1,14 @@
 import { Async } from '../Async.js'
 import { Breadcrumb, at } from '../Breadcrumb.js'
 import { Fail } from '../Fail.js'
-import { Fork, ForkContext } from '../Concurrent.js'
 import { Fx } from '../Fx.js'
 import { CapturedHandler, HandlerCapture, withHandlerContext } from '../HandlerCapture.js'
 import type { Interrupt } from '../Interrupt.js'
 import { Task } from '../Task.js'
 import { getTrace } from '../Trace.js'
 import type { Trace, TraceOptions } from '../Trace.js'
+import { Fork } from './concurrent/effects.js'
+import type { ForkContext } from './concurrent/effects.js'
 import { Semaphore } from './Semaphore.js'
 import { DisposableSet } from './disposable.js'
 import { ForkError, capturePrependTraceWithContext, captureTraceWithContext, forkFrameMetadata, originOfUnhandledFail, runtimeContextOfEffect, traceUnhandledFail, traceWithCause } from './forkDiagnostics.js'
